refactor(impulsoetl): :truck: Renomeia recursos relativos ao CNES para SCNES.
